Wrap form fields in form tag
Had it working for a moment. Clicking the button logs the fields as an
object, but then it breaks saying that "addTodo" is not a function.
I export the whole todoLibrary, and that contains the addTodo method...

I will figure this out.

Important to commit more often, especially once I get something working.