First commit of the display branch
Playing with calling a function from a module that's called from the
original module. a click calls ui/updateDisplayDefault(), which calls
library/displayTodoListDefault, which calls ui/createTodoCard.
At this time, createTodoCard just has a console log that fires for each
todoozy in the todoList.

It seems to work okay so far. Next commit will be trying out the actual
thing that makes and shows cards.